avance

+++ src/main/java/edu/eci/pdsw/sampleprj/dao/mybatis/mappers/ClienteMapper.java
@@ -11,7 +11,7 @@ public interface ClienteMapper {
     public void insertarCliente(@Param("ct")Cliente ct);
 	public Cliente consultarCliente(@Param("idcli") int id);
 	public Cliente consultarCliente(@Param("idcli") long id);
-    public void agregarItemRentadoACliente(@Param("idc") int idc, 
+    public void agregarItemRentadoACliente(@Param("idc") long idc, 
     		@Param("idi") int idi, 
     		@Param("fini") Date fechainicio,
     		@Param("ffin") Date fechafin);
